requirements.txt
streamlit==1.11.1
numpy==1.21.6
OpenCv==4.6.0
Pillow==9.2.0
ultralytics==YOLOv5

